requirements.txt
fastapi
pydantic
sqlalchemy
langchain
langchain_ollama
python-dotenv
uvicorn
psycopg2-binary